+++ app/controllers/api/v1/settings/mobile_controller.rb
@@ -53,7 +53,7 @@ class Api::V1::Settings::MobileController < ApiController
       settings: mobile.except('updated_at'),
       updated_at: mobile['updated_at'],
       capabilities: {
-        photo_library_import: { version: 1 }
+        photo_library_import: { version: MobilePhotoLibrary::Importer::FORMAT_VERSION }
       },
       status: 'success'
     }